Update CCAssistant.js
Added comments and tried to figure out what was going on with the code. 
Lots of comparison between this code and the service desk code.
Did quite a bit of poking around in TDX Elements, looks like a name has changed that makes the code obsolete.
Hard to tell what happened since I can't see the older version of Chrome's Elements.
Can't tell if the desktop module name changed, or if 
Will take a look next week to determine what exactly those names are.